+++ castle/cms/browser/trash.py
@@ -1,8 +1,10 @@
 from castle.cms import trash
 from castle.cms.interfaces import ITrashed
+from castle.cms.constants import TRASH_LOG_KEY
 from plone import api
 from plone.app.linkintegrity.exceptions import LinkIntegrityNotificationException
 from plone.locking.interfaces import ILockable
+from zope.annotation.interfaces import IAnnotations
 from Products.Five import BrowserView
 from unidecode import unidecode
 
@@ -22,6 +24,7 @@ class TrashView(BrowserView):
             elif action == 'empty':
                 self.empty()
         self.items = self.catalog(trashed=True, sort_on='modified',
+                                  sort_order='reverse',
                                   object_provides=ITrashed.__identifier__)
         return self.index()
 
@@ -79,3 +82,10 @@ class TrashView(BrowserView):
                     'to other content on the site.',
                     self.request, type='warning')
         api.portal.show_message('Trash emptied', self.request, type='warning')
+
+    def get_empty_log(self):
+        annotations = IAnnotations(self.site)
+        if TRASH_LOG_KEY in annotations:
+            return annotations[TRASH_LOG_KEY]
+        else:
+            return "No empty operation log, the trash has not been emptied"